Add docs on how to do backports
We are going to have to maintain the govuk-frontend-2 branch for a
while, at least until supplier-frontend and admin-frontend are migrated
to govuk-frontend 3.x.x, which will probably take a couple of quarters.

This commit adds some notes on how to do a backport release, for when
we've made changes to the main branch of this repo and need those
changes in the apps using the 2.x.x series.